requirements.txt
alembic==0.9.6
asn1crypto==0.23.0
beautifulsoup4==4.6.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.1.3
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.4
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.13.3
olefile==0.44
optionaldict==0.1.1
pdfminer3k==1.3.1
Pillow==4.2.1
ply==3.10
psycopg2==2.7.3
py==1.5.2
pycparser==2.18
PyPDF2==1.26.0
pytest==3.2.5
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
requests==2.18.4
scraperwiki==0.5.1
six==1.11.0
slate==0.3
SQLAlchemy==1.1.15
urllib3==1.22
wechatpy==1.5.3
Werkzeug==0.12.2
xmltodict==0.11.0
